AB#61561 Support object fields in amsterdam schema
Objects fields that are not a relation are now 'flattened' to
fit in the relational database. So, a schema definition with the
following field:

"naamGebruik": {
  "type": "object",
  "properties": {
      "code": {
	  "type": "string"
      },
      "omschrijving": {
	  "type": "string"
      }
  }
}

will generate a Django model that has fields `naam_gebruik_code` and
`naam_gebruik_omschrijving`.

Furthermore, a json type field has been added, so the following field:

"soortGrootte": {
    "type": "object",
    "format": "json"
}

will generate a jsonb field in the Postgres database.

NB. the structure of this json field is not validated! So, the only
check is that the postgres database ensures that the data can be parsed
as json.

Also the mocker has been updated to correctly supply these object fields
with fake data.

The ndjson importer has been updated so that incoming ndjson records
with object fields are processed in the correct way.

The relevant tests have been changed and new tests have been added.