Fix authorisation on array fields (#442)
Authorisation on these fields was implemented by setting an "auth" with
the union of table and field scopes on the synthetic table schema for
the field's database table. That means that having any of these scopes,
rather than at least one of the table scopes and one of the field
scopes, allowed access to the array field.

The field scopes are now set on the synthetic parent field's "auth",
which is what DSO-API checks when looking for fields to omit.