Fix code to enable attributes on a relation
For the BenK datasets relations can have their own "geldigheid".
To store this extra information, a through table is needed.

This is already in place for NM relations and for
1-N relations having a temporal table as the target.

However, for 1-N relations to a non-temporal table no through table is
being created.

The condition that was in place to determine if a 1-N relation needs
a through table was based on the temporality of the target.
This condition has been changed now. The new condition is as follows.
If the relation does contains fields that are not part of the key
a through table will be defined for this relation.

An example:
The table `brk.tenaamstelling` has the following 1-N relation field:

  "isGebaseerdOpStukdeel": {
    "type": "object",
    "properties": {
      "identificatie": {
        "type": "string"
      },
      "beginGeldigheid": {
        "type": "string",
        "format": "date-time"
      }
    },
    "relation": "brk:stukdelen"
  }

The target table `brk:stukdelen` only has a single PK `identificatie`.
So, there is an extra non-key field `beginGeldigheid` on the relation.
This will define a through table for this 1-N relation.