Make export to csv/jsonlines less memory hungy
The SQLAlchemy `execute` is loading a full set of records in memory,
leading to 4Gb of mem consumption for large datasettables like
brk.kadatastraleobjecten.

The new SA 1.4 that we are using already has a nice streaming/batching
mechanisme to reduce this memory consumption.

The SA engine has been changed now to be able to use the new connection.
No failing tests, so it seems this change is backwards compatible.