AB#91240 Add extra option to permissions cli to add additional grants in the format `<table_name>:<privilege_1>[,<privilege_n>]*;<grantee_1>[,grantee_n]*`.
Be aware that this is not completely fool-proof. If not-existing
tables/prvilege/grantees are used, the script will bail out.

This options has mainly been added for Azure, where acces to the
internal dataset_* tables (used by Django) is done in the context of a
user (or the anonymous user) and the `scope_openbaar` grant is needed
for these internal tables.

Grants can be added using a `--additional-grants` flag