Improve the sql differ to be able to use all dataset definitions
The previous sql differ could only be used with datasets that included
the tables in the same file. This new version can also handle
dataset with tables in separate files.

This is done by cloning the Amsterdam schemas repo and checking
out the git references that need to be compared.

E.g. compare PR `jan/add-some-field` with `master`.