AB#101824 Add helper method to userscopes
To determine if the user an access all fields of a particular table.
This new method is needed, because on Azure the postgresql columns
can have individual scope-based access. Django collects
the full set of columns (select * from) when accessing
a FK or M2M relation. This leads to permission error
at the database level when accessing the API.

So, more specific access of columns is needed,
hence the extra helper method on userscopes.