Block deepcopy() of serializer objects, as it's very slow
A deepcopy() happened because the schema objects were assigned as
attributes to Django REST Framework fields. DRF will deepcopy all
statically declared fields to request data and parent-fields can be
assigned to the whole tree. That operation easily took more then 14sec.

This change is there to avoid such pitfall in the future, as the copy
operation is simply too slow in its present form.