Create mock data based on amsterdam schema
The create mock data command was using an outdated and incorrect dataset
to generate the mock_model_factories. The schema was being stripped
which caused several errors.

We now use generate mock data based on the actual current version of
amsterdam schema. First we make sure all mock_models are created to be
able to have foreign keys. For now we still fill the relations with None
values, so the relate_mock_data command still works.